requirements.txt
numpy
streamlit==1.12.0
pillow==9.2.0